Login + route update
the login will display as the first page that a user will see and routes to pages works very great to the way that I expected it to be. For how I made of allowing to navigate in react is installing the route for react and import browserouter and create route or something to take in value to be used for navigation.